Refactored test to correct folder
